+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file. This change
 - [frontend profile] Add browser history replace-state re-frame events
 - [object-storage profile] Upgrade object-storage.s3 dependency version
 - [core profile] Explicitly set http-kit version instead of using transitive dependencies
+- [core profile] Add thread-transactions utility function
 
 ## [0.1.12] - 2024-04-30
 - [aws profile] Correctly handle environment variables with '$'